fix(graf): show mode banner over header bar
